refactor(lotto): restructure project into MVC folders
