+++ src/main/kotlin/lotto/Controller.kt
@@ -1,4 +1,5 @@
 package lotto
+import lotto.controller.*
 
 fun main() {
     LottoMachine.start()
