+++ src/main/kotlin/lotto/model/Numbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 class Numbers(private val numbers: List<Int>) {
     init {
