+++ src/main/kotlin/lotto/model/Rank.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 enum class Rank(val countOfMatch: Int, val prize: Int) {
     FIRST(6, 2_000_000_000),
