+++ src/main/kotlin/lotto/model/Ticket.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.model
 
 class Ticket(val numbers: Numbers)
